requirements/requirements.txt
django==4.0.3
django-environ==0.8.1
django-widget-tweaks==1.4.12
djangorestframework==3.13.1
weasyprint==54.2
django-import-export==2.8.0
xhtml2pdf==0.2.7
psycopg2-binary==2.9.3